refactor: move FileArea and other components to ui folder
