+++ src/components/ui/FileArea/index.d.ts
